requirements.txt
faster-whisper>=1.0.0
websockets>=11.0.0
numpy>=1.24.0
soundfile>=0.12.0
webrtcvad>=2.0.10
google-generativeai>=0.3.0
python-dotenv>=1.0.0
notion-client>=2.0.0
